+++ crates/miden-protocol/src/transaction/kernel/memory.rs
@@ -40,9 +40,9 @@ pub type StorageSlot = u8;
 // | Padding            | 1_056         | 4                |                                        |
 // | Proc tracking      | 1_060         | 256              | 256 procedures max, 1 element each     |
 // | Num storage slots  | 1_316         | 4                |                                        |
-// | Initial slot info  | 1_320         | 1_020            | Only initialized on the native account |
-// | Active slot info   | 2_340         | 1_020            | 255 slots max, 8 elements each         |
-// | Padding            | 3_360         | 4_832            |                                        |
+// | Initial slot info  | 1_320         | 2_040            | Only initialized on the native account |
+// | Active slot info   | 3_360         | 2_040            | 255 slots max, 8 elements each         |
+// | Padding            | 5_400         | 2_792            |                                        |
 //
 // Storage slots are laid out as [[0, slot_type, slot_id_suffix, slot_id_prefix], SLOT_VALUE].
 
@@ -333,7 +333,7 @@ pub const ACCT_STORAGE_SLOT_VALUE_OFFSET: u8 = 4;
 /// the account data segment.
 ///
 /// This section contains the current values of the account storage slots.
-pub const ACCT_ACTIVE_STORAGE_SLOTS_SECTION_OFFSET: MemoryAddress = 2340;
+pub const ACCT_ACTIVE_STORAGE_SLOTS_SECTION_OFFSET: MemoryAddress = 3360;
 
 /// The memory address at which the account's active storage slots section begins in the native
 /// account.